+++ java/src/jmri/jmrix/AbstractSerialConnectionConfig.java
@@ -81,30 +81,7 @@ abstract public class AbstractSerialConnectionConfig extends AbstractConnectionC
             p.setComboBoxLastSelection(adapter.getClass().getName() + ".baud", (String) baudBox.getSelectedItem()); // NOI18N
         });
 
-        if (adapter.getSystemConnectionMemo() != null) {
-            systemPrefixField.addActionListener(e -> checkPrefixEntry(adapter));
-            systemPrefixField.addFocusListener(new FocusListener() {
-                @Override
-                public void focusLost(FocusEvent e) {
-                    checkPrefixEntry(adapter);
-                }
-
-                @Override
-                public void focusGained(FocusEvent e) {
-                }
-            });
-            connectionNameField.addActionListener(e -> checkNameEntry(adapter));
-            connectionNameField.addFocusListener(new FocusListener() {
-                @Override
-                public void focusLost(FocusEvent e) {
-                    checkNameEntry(adapter);
-                }
-
-                @Override
-                public void focusGained(FocusEvent e) {
-                }
-            });
-        }
+        addNameEntryCheckers(adapter);
 
         portBox.addFocusListener(new FocusListener() {
             @Override
